Delete useCategories, ad queries file
